+++ driver/service.go
@@ -18,7 +18,7 @@ func NewService(r Repository, l *logging.Logging) *Service {
 	}
 }
 
-func (s *Service) GetByDriverId(ctx context.Context, id int) (*Driver, error) {
+func (s *Service) GetByDriverId(ctx context.Context, id int64) (*Driver, error) {
 	driver, err := s.repo.GetById(ctx, id)
 	if err != nil {
 		return nil, err
@@ -27,7 +27,7 @@ func (s *Service) GetByDriverId(ctx context.Context, id int) (*Driver, error) {
 	return driver, nil
 }
 
-func (s *Service) GetByUserId(ctx context.Context, userId int) (*Driver, error) {
+func (s *Service) GetByUserId(ctx context.Context, userId int64) (*Driver, error) {
 	driver, err := s.repo.GetByUserId(ctx, userId)
 	if err != nil {
 		return nil, err
@@ -36,7 +36,7 @@ func (s *Service) GetByUserId(ctx context.Context, userId int) (*Driver, error) 
 	return driver, nil
 }
 
-func (s *Service) GetByIdWithEagerLoading(ctx context.Context, id int) (*Driver, error) {
+func (s *Service) GetByIdWithEagerLoading(ctx context.Context, id int64) (*Driver, error) {
 	driver, err := s.repo.GetByIdWithEagerLoading(ctx, id)
 	if err != nil {
 		return driver, nil
@@ -45,7 +45,7 @@ func (s *Service) GetByIdWithEagerLoading(ctx context.Context, id int) (*Driver,
 	return driver, nil
 }
 
-func (s *Service) GetByUserIdWithEagerLoading(ctx context.Context, userId int) (*Driver, error) {
+func (s *Service) GetByUserIdWithEagerLoading(ctx context.Context, userId int64) (*Driver, error) {
 	driver, err := s.repo.GetByUserIdWithEagerLoading(ctx, userId)
 	if err != nil {
 		return driver, nil
@@ -72,7 +72,7 @@ func (s *Service) ListWithEagerLoading(ctx context.Context, specification *Drive
 	return drivers, nil
 }
 
-func (s *Service) Create(ctx context.Context, d *Driver) (int, error) {
+func (s *Service) Create(ctx context.Context, d *Driver) (int64, error) {
 	driverID, err := s.repo.Create(ctx, d)
 	if err != nil {
 		return 0, err
@@ -85,6 +85,6 @@ func (s *Service) Update(ctx context.Context, d *Driver) error {
 	return s.repo.Update(ctx, d)
 }
 
-func (s *Service) Delete(ctx context.Context, id int) error {
+func (s *Service) Delete(ctx context.Context, id int64) error {
 	return s.repo.Delete(ctx, id)
 }
